feat: Add SuccessResponse object
