+++ src/config/middleware.py
@@ -2,12 +2,12 @@ from datetime import datetime
 from fastapi.responses import JSONResponse
 from starlette.requests import Request
 from src.auth.util.jwt import parse_token
-from src.exceptions.definitions import (
+from src.response.error_definitions import (
     BaseAppException,
     InvalidJwtToken,
     JwtTokenNotFound,
 )
-from src.exceptions.schemas import ErrorResponse
+from src.response.schemas import ErrorResponse
 from src.config.logger_config import setup_logger, add_daily_file_handler
 
 logger = setup_logger(__name__)
