--- src/exceptions/schemas.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-from datetime import datetime
-
-class ErrorResponse(BaseModel):
-    message: str
-    timestamp: str
