+++ src/user/repository.py
@@ -2,7 +2,7 @@ from sqlalchemy import select
 from sqlalchemy.orm import Session
 from sqlalchemy.orm.exc import NoResultFound
 from sqlalchemy.exc import SQLAlchemyError
-from src.exceptions.definitions import SQLError
+from src.response.error_definitions import SQLError
 from src.config.logger_config import add_daily_file_handler, setup_logger
 from src.models import User
